procs/processamento/src/pipeline/requirements.txt
MySQL-python==1.2.5
PyStemmer==1.3.0
anyjson==0.3.3
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
certifi==14.05.14
-e git+https://github.com/decultured/Python-Language-Detector.git@cf77eb42ee053e7e21ac9dd45b0ca68e75190bb2#egg=languageIdentifier-dev
pika==0.9.13
pyes==0.99.2
pymongo==2.7.1
pynsq==0.6.4
pytz==2014.4
redis==2.10.2
requests==2.4.1
simplejson==3.5.2
six==1.7.3
solrpy==0.9.6
tornado==4.0.2
urllib3==1.9
wsgiref==0.1.2
